Move apiTypes.ts and rename
