+++ src/test/common.ts
@@ -10,7 +10,7 @@ import * as glob from 'glob';
 import * as path from 'path';
 import { coerce, SemVer } from 'semver';
 import { ConfigurationTarget, Event, TextDocument, Uri } from 'vscode';
-import type { IExtensionApi } from '../client/apiTypes';
+import type { PythonExtension } from '../client/api/apiTypes';
 import { IProcessService } from '../client/common/process/types';
 import { IDisposable } from '../client/common/types';
 import { IServiceContainer, IServiceManager } from '../client/ioc/types';
@@ -438,7 +438,7 @@ export async function isPythonVersion(...versions: string[]): Promise<boolean> {
     }
 }
 
-export interface IExtensionTestApi extends IExtensionApi, ProposedExtensionAPI {
+export interface IExtensionTestApi extends PythonExtension, ProposedExtensionAPI {
     serviceContainer: IServiceContainer;
     serviceManager: IServiceManager;
 }
