+++ src/test/initialize.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import * as vscode from 'vscode';
-import type { IExtensionApi } from '../client/apiTypes';
+import type { PythonExtension } from '../client/api/apiTypes';
 import {
     clearPythonPathInWorkspaceFolder,
     IExtensionTestApi,
@@ -42,7 +42,7 @@ export async function initialize(): Promise<IExtensionTestApi> {
     return (api as any) as IExtensionTestApi;
 }
 export async function activateExtension() {
-    const extension = vscode.extensions.getExtension<IExtensionApi>(PVSC_EXTENSION_ID_FOR_TESTS)!;
+    const extension = vscode.extensions.getExtension<PythonExtension>(PVSC_EXTENSION_ID_FOR_TESTS)!;
     const api = await extension.activate();
     // Wait until its ready to use.
     await api.ready;
